x86/tdx: Use u64 return type for tdx_hcall_get_quote()
Currently, tdx_hcall_get_quote() function uses incorrect size (int) for
the return error code. Since GetQuote hypercall returns 64 bit error
code, use u64 as return type.

Since we are using standard hypercall type, use _tdx_hypercall() wrapper
instead of initializing args and using __tdx_hypercall().

+++ arch/x86/include/asm/tdx.h
@@ -58,7 +58,7 @@ int tdx_register_event_irq_cb(tdx_event_irq_cb_t handler, void *data);
 
 int tdx_unregister_event_irq_cb(tdx_event_irq_cb_t handler, void *data);
 
-int tdx_hcall_get_quote(u8 *buf, size_t size);
+u64 tdx_hcall_get_quote(u8 *buf, size_t size);
 
 #else
